working on if statement checking remains file if exists, but still need some work.. wrong statement.

+++ vuadek.py
@@ -13,13 +13,15 @@ zm_fl_remains = zm_pth_workdir+"remains"
 
 pathname = os.path.dirname(sys.argv[1])
 
-scan_result = subprocess.Popen(["uade123", "--scan", os.path.abspath(pathname)],stdout=subprocess.PIPE)
+if not os.path.isfile(zm_fl_remains):
+	print('nie istnieje')
+	scan_result = subprocess.Popen(["uade123", "--scan", os.path.abspath(pathname)],stdout=subprocess.PIPE)
+	with open(zm_fl_remains, 'w') as f:
+		for line in scan_result.stdout:
+			f.write(line.decode('utf-8'))		
+	f.closed
 
-with open(zm_fl_remains, 'w') as f:
-	for line in scan_result.stdout:
-		f.write(line.decode('utf-8'))		
-f.closed
-
+print('istnieje')
 with open(zm_fl_remains, 'r') as f:
 	zm_input = [line.rstrip('\n') for line in f]
 	for item in zm_input:
